pyproject.toml
[tool.poetry]
name = "usps-client"
version = "0.2"
description = "An unofficial client for the USPS Web Tools APIs"
authors = ["macro1 <macromicah@gmail.com>"]
license = "ISC"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
lxml = {version = "*", optional = true}
urllib3 = {version = ">=1.9,<2", extras = ["secure"]}
attrs = ">=17"
inflection = "*"

[tool.poetry.extras]
lxml = ["lxml"]

[tool.poetry.dev-dependencies]
black = "*"
mypy = "*"
pytest = "^6.2"
types-attrs = "^19.1.0"
types-certifi = "^2021.10.8"
lxml-stubs = { git = "https://github.com/lxml/lxml-stubs.git", branch = "master" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = "src,tests"
disallow_any_explicit = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "usps_client.base_model"
disallow_any_explicit = false

[[tool.mypy.overrides]]
# Ignore libraries that are not compatible
module = [
    "urllib3"
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# Don't require annotations in tests
module = [
    "integration.*",
    "unit.*"
]
disallow_untyped_defs = false


[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::ImportWarning"
]
